this is it
problem is that query is scaled to 0 and it takes some time to bring it
up. k6 for some reason manages to tear down already and abort on failure
happens.

Don't see the timeout for handshake or whatever (maybe there is one?).
If not found, we need warming up rounds (we need them anyway though)
where we don't care about number of requests.