can this work
Reminder: we need to assure that upload starts only when create is
finished.
So init containers are the best option as container is not considered up
while these are not complete.
It can probably be done in the same container, but we need logs.
Logs from init container are available only through machinations like
that.