Suspicion: there is a deadlock in fileshare. Lets try not running stuff
Apparently upload does start running before create finishes. So "init
containers" help not.
With addition to finishing too fast, it might be that they both request
same fileshare, and maybe something doesn't play along nicely.

So if that works, both containers must be init, or we need to wait for
Terminated state (we need that in any case to run the tests)