Revert "lets try emptydir and one job again"
This reverts commit 6e023eeaaf9dfc77af0179acb540fdcba1ab0429.

There seems to be no benefit. Create is a bit slower as it is saved on
drive, upload maybe faster because it doesn't have to wait on network.
But it is still 3 minutes on create vs 15 minutes on upload, so maybe
emptyDir is actually a network drive too.
Probably won't bring significant benefits. Also we are limited by 50 GB
in that case.